Add in-URL parameters for GET operations
We want to support in-URL parameters for GET operations, as opposed
to the currently used in-body parameters. This is a more RESTful way
of doing things.

This commit adds in-URL parameters in parallel with in-body params,
to support older DLRN API releases. In the future, we will switch
to in-URL parameters only.

TG-2650

Change-Id: I8371251e701539d690ffb28118689d297341ef75