Add special case for "--extended-hash=None" in cmdline
With the added support for the --extended-hash parameter to some
calls, we have had some confusion around the behavior when the
paremeter needs to be "None". Just omitting the switch was enough,
however we want to support the "--extended-hash None" option as
well.

This commit adds a special treatment for that string, and turns
the "None" string into a Python None value.

Change-Id: I710c0e36e07ea66f894e15e24903c368943d7ebc