Allow specifying a custom CA bundle file
In some cases, we may want to keep SSL certificate verification set,
but use our own CA bundle file when connecting to an SSL -enabled
API server.

This commit adds a check for a new environment variable, SSL_CA_BUNDLE.
When set, it will instruct the client to use that file to verify SSL
certificates, instead of the default one provided by certifi.

Change-Id: I7f26798d7e1bd91493a12ca6f7e4c6f6fde57482