Add Kerberos auth method.
Adding Kerberos Authentication for consuming kerberos-protected endpoints.
Default Auth protocol is Basic, but now is possible to select Kerberos and the server principal name as well.
This returns a valid token that the server would be able to decrypt if the principal exists in the Kerberos' database.

Gssapi dependency set up as Optional.

Change-Id: I736718b2ab8ca4999263c3f93f74be43deafeb1f